+++ classes/WpMatomo/Admin/SystemReport.php
@@ -20,8 +20,10 @@ use WpMatomo\Capabilities;
 use WpMatomo\Paths;
 use WpMatomo\ScheduledTasks;
 use WpMatomo\Settings;
+use WpMatomo\Site;
 use WpMatomo\Site\Sync as SiteSync;
 use WpMatomo\User\Sync as UserSync;
+use function DI\value;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // if accessed directly
@@ -106,8 +108,8 @@ class SystemReport {
 			$tables = array(
 				array( 'title' => 'Matomo', 'rows' => $this->get_matomo_info(), 'has_comments' => true ),
 				array( 'title' => 'WordPress', 'rows' => $this->get_wordpress_info() ),
-				array( 'title' => 'WordPress Plugins', 'rows' => $this->get_plugins_info() ),
-				array( 'title' => 'Server', 'rows' => $this->get_server_info() ),
+				array( 'title' => 'WordPress Plugins', 'rows' => $this->get_plugins_info(), 'has_comments' => true ),
+				array( 'title' => 'Server', 'rows' => $this->get_server_info(), 'has_comments' => true ),
 			);
 		}
 
@@ -120,18 +122,18 @@ class SystemReport {
 		$file_writable = is_writable( $path_to_check );
 		$comment       = '"' . $path_to_check . '"';
 		if ( ! $file_exists ) {
-			$comment .= ' ' . $title . ' does not exist.';
+			$comment .= sprintf( __( '%s does not exist.', 'matomo' ), $title );
 		}
 		if ( ! $file_readable ) {
-			$comment .= ' ' . $title . ' is not readable.';
+			$comment .= sprintf( __( '%s is not readable.', 'matomo' ), $title );
 		}
 		if ( ! $file_writable ) {
-			$comment .= ' ' . $title . ' is not writable.';
+			$comment .= sprintf( __( '%s is not writable.', 'matomo' ), $title );
 		}
 
 		$rows[] = array(
-			'name'    => $title . ' exists and is writable',
-			'value'   => $file_exists && $file_readable && $file_writable ? 'Yes' : 'No',
+			'name'    => sprintf( __( '%s exists and is writable.', 'matomo' ), $title ),
+			'value'   => $file_exists && $file_readable && $file_writable ? __( 'Yes' ) : __( 'No' ),
 			'comment' => $comment
 		);
 
@@ -143,7 +145,11 @@ class SystemReport {
 
 		$plugin_data = get_plugin_data( MATOMO_ANALYTICS_FILE, $markup = false, $translate = false );
 
-		$rows[] = array( 'name' => 'Matomo WordPress Version', 'value' => $plugin_data['Version'], 'comment' => '' );
+		$rows[] = array(
+			'name'    => __( 'Matomo Plugin Version', 'matomo' ),
+			'value'   => $plugin_data['Version'],
+			'comment' => ''
+		);
 
 		$paths            = new Paths();
 		$upload_dir       = $paths->get_upload_base_dir();
@@ -154,7 +160,7 @@ class SystemReport {
 		$rows              = $this->check_file_exists_and_writable( $rows, $path_tracker_file, 'JS Tracker' );
 
 		$rows[] = array(
-			'name'    => 'Plugin directories',
+			'name'    => __( 'Plugin directories', 'matomo' ),
 			'value'   => ! empty( $GLOBALS['MATOMO_PLUGIN_DIRS'] ) ? 'Yes' : 'No',
 			'comment' => ! empty( $GLOBALS['MATOMO_PLUGIN_DIRS'] ) ? json_encode( $GLOBALS['MATOMO_PLUGIN_DIRS'] ) : ''
 		);
@@ -162,7 +168,7 @@ class SystemReport {
 		$tmp_dir = $paths->get_tmp_dir();
 
 		$rows[] = array(
-			'name'    => 'Tmp directory writable',
+			'name'    => __( 'Tmp directory writable', 'matomo' ),
 			'value'   => is_writable( $tmp_dir ),
 			'comment' => $tmp_dir
 		);
@@ -175,7 +181,7 @@ class SystemReport {
 
 		} catch ( \Exception $e ) {
 			$rows[] = array(
-				'name'    => 'Matomo System Check',
+				'name'    => __( 'Matomo System Check', 'matomo' ),
 				'value'   => 'Failed to run, please open the system check in Matomo',
 				'comment' => ''
 			);
@@ -184,8 +190,17 @@ class SystemReport {
 		}
 
 		$rows[] = array(
-			'name'    => 'Matomo Version',
+			'name'    => __( 'Matomo Version', 'matomo' ),
 			'value'   => \Piwik\Version::VERSION,
+			'comment' => ''
+		);
+
+		$site = new Site();
+		$idsite = $site->get_current_matomo_site_id();
+
+		$rows[] = array(
+			'name'    => __( 'Matomo Blog idSite', 'matomo' ),
+			'value'   => $idsite,
 			'comment' => ''
 		);
 
@@ -222,46 +237,46 @@ class SystemReport {
 		);
 
 		$scheduled_tasks = new ScheduledTasks( $this->settings );
-		$all_events = $scheduled_tasks->get_all_events();
+		$all_events      = $scheduled_tasks->get_all_events();
 
 		$rows[] = array(
-			'name'    => 'Server time',
-			'value'   => $this->convert_time_to_date(time(), false),
+			'name'    => __( 'Server time', 'matomo' ),
+			'value'   => $this->convert_time_to_date( time(), false ),
 			'comment' => ''
 		);
 
 		$rows[] = array(
-			'name'    => 'Blog time',
-			'value'   => $this->convert_time_to_date(time(), true),
-			'comment' => 'Below dates are shown in blog timezone'
+			'name'    => __( 'Blog time', 'matomo' ),
+			'value'   => $this->convert_time_to_date( time(), true ),
+			'comment' => __( 'Below dates are shown in blog timezone', 'matomo' )
 		);
 
-		foreach ($all_events as $event_name => $event_config) {
-			$last_run_before = $scheduled_tasks->get_last_time_before_cron($event_name);
-			$last_run_after = $scheduled_tasks->get_last_time_after_cron($event_name);
+		foreach ( $all_events as $event_name => $event_config ) {
+			$last_run_before = $scheduled_tasks->get_last_time_before_cron( $event_name );
+			$last_run_after  = $scheduled_tasks->get_last_time_after_cron( $event_name );
 
-			$next_scheduled = wp_next_scheduled($event_name);
+			$next_scheduled = wp_next_scheduled( $event_name );
 
-			$comment  = ' Last started: ' . $this->convert_time_to_date($last_run_before, true, true) . '.';
-			$comment .= ' Last ended: ' . $this->convert_time_to_date($last_run_after, true, true) . '.';
+			$comment = ' Last started: ' . $this->convert_time_to_date( $last_run_before, true, true ) . '.';
+			$comment .= ' Last ended: ' . $this->convert_time_to_date( $last_run_after, true, true ) . '.';
 			$comment .= ' Interval: ' . $event_config['interval'];
 
 			$rows[] = array(
 				'name'    => $event_config['name'],
-				'value'   => 'Next run: ' . $this->convert_time_to_date($next_scheduled, true, true),
+				'value'   => 'Next run: ' . $this->convert_time_to_date( $next_scheduled, true, true ),
 				'comment' => $comment
 			);
 
 		}
 
 		$rows[] = array(
-			'section' => 'Mandatory checks',
+			'section' => __( 'Mandatory checks', 'matomo' ),
 		);
 
 		$rows = $this->add_diagnostic_results( $rows, $report->getMandatoryDiagnosticResults() );
 
 		$rows[] = array(
-			'section' => 'Optional checks',
+			'section' => __( 'Optional checks', 'matomo' ),
 		);
 		$rows   = $this->add_diagnostic_results( $rows, $report->getOptionalDiagnosticResults() );
 
@@ -276,20 +291,19 @@ class SystemReport {
 		return $rows;
 	}
 
-	private function convert_time_to_date($time, $in_blog_timezone, $print_diff = false)
-	{
-		if (empty($time)) {
-			return 'Unknown';
+	private function convert_time_to_date( $time, $in_blog_timezone, $print_diff = false ) {
+		if ( empty( $time ) ) {
+			return __( 'Unknown', 'matomo' );
 		}
 
 		$date = date( 'Y-m-d H:i:s', $time );
 
-		if ($in_blog_timezone) {
+		if ( $in_blog_timezone ) {
 			$date = get_date_from_gmt( $date, 'Y-m-d H:i:s' );
 		}
 
-		if ($print_diff && class_exists('\Piwik\MetricsFormatter')) {
-			$date .= ' (' . MetricsFormatter::getPrettyTimeFromSeconds($time - time(), true, false, true ) . ')';
+		if ( $print_diff && class_exists( '\Piwik\MetricsFormatter' ) ) {
+			$date .= ' (' . MetricsFormatter::getPrettyTimeFromSeconds( $time - time(), true, false, true ) . ')';
 		}
 
 		return $date;
@@ -354,11 +368,14 @@ class SystemReport {
 		$rows[] = array( 'name' => 'MySQL Version', 'value' => ! empty( $wpdb->is_mysql ) ? $wpdb->db_version() : '' );
 		$rows[] = array( 'name' => 'Timezone', 'value' => date_default_timezone_get() );
 		$rows[] = array( 'name' => 'Locale', 'value' => get_locale() );
-		$rows[] = array( 'name' => 'Memory Limit', 'value' => max( WP_MEMORY_LIMIT, @ini_get( 'memory_limit' ) ) );
+		$rows[] = array( 'name' => 'Memory Limit', 'value' => max( WP_MEMORY_LIMIT, @ini_get( 'memory_limit' ) ), 'comment' => 'At least 128MB recommended. Depending on your traffic 256MB or more may be needed.' );
 		$rows[] = array( 'name' => 'Time', 'value' => time() );
 
 		$rows[] = array( 'name' => 'Mysqli Connect', 'value' => function_exists( 'mysqli_connect' ) );
-		$rows[] = array( 'name' => 'Force MySQL over Mysqli', 'value' => defined('WP_USE_EXT_MYSQL') && WP_USE_EXT_MYSQL );
+		$rows[] = array(
+			'name'  => 'Force MySQL over Mysqli',
+			'value' => defined( 'WP_USE_EXT_MYSQL' ) && WP_USE_EXT_MYSQL
+		);
 
 		$rows[] = array( 'name' => 'Max Execution Time', 'value' => ini_get( 'max_execution_time' ) );
 		$rows[] = array( 'name' => 'Max Post Size', 'value' => ini_get( 'post_max_size' ) );
@@ -385,7 +402,11 @@ class SystemReport {
 			);
 
 			foreach ( $mu_plugins as $mu_pin ) {
-				$rows[] = array( 'name' => $mu_pin['Name'], 'value' => $mu_pin['Version'] );
+				$comment = '';
+				if ( ! empty( $plugin['Network'] ) ) {
+					$comment = 'Network enabled';
+				}
+				$rows[] = array( 'name' => $mu_pin['Name'], 'value' => $mu_pin['Version'], 'comment' => $comment );
 			}
 
 			$rows[] = array(
@@ -396,8 +417,22 @@ class SystemReport {
 		$plugins = get_plugins();
 
 		foreach ( $plugins as $plugin ) {
-			$rows[] = array( 'name' => $plugin['Name'], 'value' => $plugin['Version'] );
+			$comment = '';
+			if ( ! empty( $plugin['Network'] ) ) {
+				$comment = 'Network enabled';
+			}
+			$rows[] = array( 'name' => $plugin['Name'], 'value' => $plugin['Version'], 'comment' => $comment );
 		}
+
+		$active_plugins = get_option( 'active_plugins', array() );
+		if ( ! empty( $active_plugins ) && is_array( $active_plugins ) ) {
+			$rows[] = array(
+				'name'    => 'Active Plugins',
+				'value'   => count( $active_plugins ),
+				'comment' => implode( ', ', $active_plugins )
+			);
+		}
+
 
 		return $rows;
 	}