+++ classes/WpMatomo/Admin/views/get_started.php
@@ -29,38 +29,39 @@ if ( empty( $show_this_page ) ) {
 <div class="wrap">
     <div id="icon-plugins" class="icon32"></div>
 
-    <h1>Start getting a full picture of your visitors</h1>
+    <h1><?php _e('Start getting a full picture of your visitors', 'matomo') ?></h1>
 
 	<?php if ( $was_updated ) {
 		include 'update_notice_clear_cache.php';
 	} ?>
 
 	<?php if ( $settings->is_tracking_enabled() ) { ?>
-        <h2>1. Tracking is enabled <span class="dashicons dashicons-yes"></span></h2>
+        <h2>1. <?php _e('Tracking is enabled', 'matomo') ?> <span class="dashicons dashicons-yes"></span></h2>
         <p><a href="<?php echo AdminSettings::make_url( AdminSettings::TAB_TRACKING ) ?>">Click here to configure your
                 tracking code.</a></p>
 
 	<?php } else { ?>
-        <h2>1. Enable tracking</h2>
+        <h2>1. <?php _e('Enable tracking', 'matomo'); ?></h2>
 
-        <form method="post">Tracking is currently disabled <?php wp_nonce_field( GetStarted::NONCE_NAME ); ?>
+        <form method="post"><?php _e('Tracking is currently disabled', 'matomo'); ?> <?php wp_nonce_field( GetStarted::NONCE_NAME ); ?>
             <input type="hidden" name="<?php echo GetStarted::FORM_NAME ?>[track_mode]"
                    value="<?php echo esc_attr( TrackingSettings::TRACK_MODE_DEFAULT ) ?>">
-            <input type="submit" class="button-primary" value="Enable tracking now">
+            <input type="submit" class="button-primary" value="<?php _e( 'Enable tracking now', 'matomo' ); ?>">
         </form>
 	<?php } ?>
 
-    <h2>2. Update your privacy page</h2>
-    Give your users the chance to opt-out of tracking by adding the shortcode <code>[matomo_opt_out]</code> to your
-    privacy page. You can <a href="<?php echo AdminSettings::make_url( AdminSettings::TAB_PRIVACY ) ?>">tweak the
-        opt-out to your liking see the Privacy Settings</a>.
-    You may also need to mention that you are using Matomo Analytics on your website.
-    <h2>3. Done</h2>
+    <h2>2. <?php _e( 'Update your privacy page', 'matomo' ); ?></h2>
+
+    <?php echo sprintf( __( 'Give your users the chance to opt-out of tracking by adding the shortcode %1$s to your privacy page. You can %2$stweak the opt-out to your liking - see the Privacy Settings%3$s.', 'matomo' ), '<code>[matomo_opt_out]</code>', '<a href="' . AdminSettings::make_url( AdminSettings::TAB_PRIVACY ) .'">', '</a>' ); ?>
+
+    <?php _e( 'You may also need to mention that you are using Matomo Analytics on your website.', 'matomo' ); ?>
+
+    <h2>3. <?php _e( 'Done', 'matomo' ); ?></h2>
     <form method="post">
 		<?php wp_nonce_field( GetStarted::NONCE_NAME ); ?>
         <input type="hidden" name="<?php echo GetStarted::FORM_NAME ?>[show_get_started_page]"
                value="no">
-        <input type="submit" class="button-primary" value="Don't show this page anymore">
+        <input type="submit" class="button-primary" value="<?php _e( 'Don\'t show this page anymore', 'matomo'); ?>">
     </form>
     <p>
         <br/>