+++ classes/WpMatomo/Admin/views/info_bug_report.php
@@ -1,9 +1,10 @@
-<h2>Do you have a bug to report or a feature request?</h2>
-<p>Please read the recommendations on writing a good <a target="_blank" rel="noreferrer noopener"
-                                                        href="https://developer.matomo.org/guides/core-team-workflow#submitting-a-bug-report">bug
-		report</a> or <a target="_blank" rel="noreferrer noopener"
-                         href="https://developer.matomo.org/guides/core-team-workflow#submitting-a-feature-request">feature
-		request</a>. Then register or login on <a target="_blank" rel="noreferrer noopener"
-                                                  href="https://github.com/matomo-org/wp-matomo/issues">our issue
-		tracker</a> and create a <a target="_blank" rel="noreferrer noopener"
-                                    href="https://github.com/matomo-org/wp-matomo/issues/new">new issue</a>.</p>
+<h2><?php _e( 'Do you have a bug to report or a feature request?', 'matomo' ); ?></h2>
+<p><?php echo sprintf( __( 'Please read the recommendations on writing a good %1$sbug report%2$s or %3$sfeature request%4$s. Then register or login to %5$sour issue tracker%6$s and create a %7$snew issue%8$s.', 'matomo' ),
+		'<a target="_blank" rel="noreferrer noopener" href="https://developer.matomo.org/guides/core-team-workflow#submitting-a-bug-report">',
+        '</a>',
+		'<a target="_blank" rel="noreferrer noopener" href="https://developer.matomo.org/guides/core-team-workflow#submitting-a-feature-request">',
+        '</a>',
+        '<a target="_blank" rel="noreferrer noopener" href="https://github.com/matomo-org/wp-matomo/issues">',
+        '</a>',
+        '<a target="_blank" rel="noreferrer noopener" href="https://github.com/matomo-org/wp-matomo/issues/new">',
+        '</a>' ); ?></p>